Sidebar: let sidebar take 100% of the height and 20% of the width and resize accordingly
