+++ Plans/sprint_7(in progress)/Plan.md
@@ -4,9 +4,7 @@
     - bfs start from node 1 so if the user's input doesn't have node 1 it won't work 
     - the bfs act strange when we create an edge from node 2 to 1 (not from 1 to 2)
 -----
-- drag need to be only from the header
-- the nodeSize range need to be fixed
-- node count need to be fixed
+
 
 ## add a header to the site
 - be creative
