+++ js/sidebar/buttons/Clear.js
