+++ js/sidebar/buttons/Delete.js
