+++ js/sidebar/buttons/GraphSetting.js
