+++ js/sidebar/buttons/color-picker.js
